@@ -9,12 +9,12 @@
     "config": {
         "sort-packages": true
     },
-    "version": "100.3.4",
+    "version": "100.3.5",
     "require": {
         "magento/framework": "102.0.*",
         "magento/magento-composer-installer": "*",
         "magento/module-store": "101.0.*",
-        "php": "~7.1.3||~7.2.0||~7.3.0"
+        "php": "~7.3.0||~7.4.0"
     },
     "autoload": {
         "files": [
